fix endpoint form
