+++ pwa/src/pages/endpoints/[id].tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import Layout from "../../components/common/layout";
-import AttributeTable from "../../components/attributes/attributeTable";
 import { Tabs } from "@conductionnl/nl-design-system/lib/Tabs/src/tabs";
-import EndpointForm from "../../components/endpoints/endpointForm";
+import EndpointForm from "../../components/endpoints/form";
 import HandlerTable from "../../components/handlers/handlerTable";
 
 const IndexPage = (props) => {
